requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
nltk
scikit-learn
spacy
sqlalchemy
python-multipart
python-dotenv
pydantic-settings
pdfplumber
fuzzywuzzy
python-Levenshtein
python-dateutil
httpx
aiosqlite
psycopg2-binary
asyncpg
sentence-transformers



